backend/requirements.txt
asgiref==3.10.0
certifi==2025.10.5
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.4.4
daphne==4.1.0
Django==5.2.7
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
dotenv==0.9.9
gunicorn==23.0.0
idna==3.11
packaging==25.0
pillow==12.0.0
python-barcode==0.16.1
python-dotenv==1.1.1
redis==5.0.1
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.11.0
